third_party/blink/web_tests/http/tests/websocket/workers/multiple-subprotocols-expected.txt

Test WebSocket subprotocol negotiation in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: ws.protocol is equal to ""
INFO: Connected
PASS PASS: ws.protocol is equal to "superchat"
INFO: Received: superchat
PASS PASS: receivedMessage is equal to "superchat"
INFO: Closed
PASS PASS: ws.protocol is equal to "superchat"
PASS PASS: closeEvent.wasClean is true
INFO: Exited onclose handler
PASS PASS: ws.protocol is equal to "superchat"
DONE
PASS successfullyParsed is true

TEST COMPLETE
